// src/cyfs-stack-test-typescript/cyfs_node/cyfs_node.d.ts
export class DecAppId extends NamedObjectId<DecAppDescContent, DecAppBodyContent> {
    constructor(id: ObjectId);
    static default(): DecAppId;
    static from_base_58(s: string): BuckyResult<DecAppId>;
    static try_from_object_id(id: ObjectId): BuckyResult<DecAppId>;
}
export class DecAppIdDecoder extends NamedObjectIdDecoder<DecAppDescContent, DecAppBodyContent> {
    constructor();
}
// ...
export class BuckyFixedBuffer implements RawEncode {
    buffer: Uint8Array;
    constructor(buffer: Uint8Array);
    raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
    value(): Uint8Array;
}
export class BuckyFixedBufferDecoder implements RawDecode<BuckyFixedBuffer> {
    size: number;
    constructor(size: number);
    raw_decode(buf: Uint8Array, ctx?: any): BuckyResult<[BuckyFixedBuffer, Uint8Array]>;
}
export class BuckyBuffer implements RawEncode {
    buffer: Uint8Array;
    size: BuckySize;
    constructor(buf: Uint8Array);
    value(): Uint8Array;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class BuckyBufferDecoder implements RawDecode<BuckyBuffer> {
    raw_decode(buf: Uint8Array): BuckyResult<[BuckyBuffer, Uint8Array]>;
}
// ...
export class BuckyHashMap<K extends RawEncode & Compareable<K>, V extends RawEncode> implements RawEncode {
    constructor();
    get size(): number;
    clear(): void;
    delete(key: K): void;
    has(key: K): boolean;
    set(key: K, v: V): void;
    get(key: K): V | undefined;
    keys(): IterableIterator<K>;
    values(): IterableIterator<V>;
    entries(): IterableIterator<[K, V]>;
    forEach(callback: (value: V, key: K, map: BuckyHashMap<K, V>) => void): void;
    to<K1, V1>(ke: (k: K) => K1, ve: (v: V) => V1): Map<K1, V1>;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class BuckyHashMapDecoder<K extends RawEncode & Compareable<K>, V extends RawEncode> implements RawDecode<BuckyHashMap<K, V>> {
    readonly size_decoder: BuckySizeDecoder;
    readonly key_decoder: RawDecode<K>;
    readonly value_decoder: RawDecode<V>;
    constructor(key_decoder: RawDecode<K>, value_decoder: RawDecode<V>);
    raw_decode(buf: Uint8Array): BuckyResult<[BuckyHashMap<K, V>, Uint8Array]>;
}
// ...
export class BuckyHashSet<T extends RawEncode & Compareable<T>> implements RawEncode {
    constructor();
    get size(): number;
    add(v: T): BuckyHashSet<T>;
    clear(): void;
    delete(v: T): boolean;
    has(v: T): boolean;
    keys(): IterableIterator<T>;
    values(): IterableIterator<T>;
    entries(): IterableIterator<[T, T]>;
    forEach(callback: (value: T, value2: T, set: BuckyHashSet<T>) => void): void;
    to<K1>(ke: (k: T) => K1): Set<K1>;
    array(): T[];
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class BuckyHashSetDecoder<T extends RawEncode & Compareable<T>> implements RawDecode<BuckyHashSet<T>> {
    readonly decoder: RawDecode<T>;
    readonly size_decoder: BuckySizeDecoder;
    constructor(decoder: RawDecode<T>);
    raw_decode(buf: Uint8Array): BuckyResult<[BuckyHashSet<T>, Uint8Array]>;
}
// ...
export class BuckyNumber implements RawEncode {
        readonly type: Int;
        val: JSBI;
        constructor(type: Int, val: JSBI | number | string);
        value(): JSBI;
        raw_measure(): BuckyResult<number>;
        raw_encode(buf: Uint8Array, littleEndian?: boolean): BuckyResult<Uint8Array>;
        toBigInt(): JSBI;
        toNumber(): number;
}
export class BuckyNumberDecoder implements RawDecode<BuckyNumber> {
        readonly type: Int;
        constructor(type: Int);
        raw_decode(buf: Uint8Array, littleEndian?: boolean): BuckyResult<[BuckyNumber, Uint8Array]>;
}
// ...
export class BuckyString implements RawEncode, Compareable<BuckyString> {
    constructor(value: string);
    hashCode(): symbol;
    equals(other: BuckyString): boolean;
    value(): string;
    toString(): string;
    raw_measure(ctx?: any): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
export class BuckyStringDecoder implements RawDecode<BuckyString> {
    raw_decode(buf: Uint8Array, ctx?: any): BuckyResult<[BuckyString, Uint8Array]>;
}
// ...
export class BuckySize implements RawEncode {
    constructor(size: number | JSBI | string);
    raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
}
export class BuckySizeDecoder implements RawDecode<number> {
    raw_decode(buf: Uint8Array, ctx?: any): BuckyResult<[number, Uint8Array]>;
}

export class BuckyTuple implements RawEncode {
    members: RawEncode[];
    constructor(members: RawEncode[]);
    raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
    index<T extends RawEncode>(index: number): T;
}
// ...
export enum Protocol {
    Unk = 0,
    Tcp = 1,
    Udp = 2
}
// ...
export class Endpoint implements RawEncode {
    is_static_wan: boolean;
    protocol: Protocol;
    addr: SocketAddr;
    constructor(is_static_wan: boolean, protocol: Protocol, addr: SocketAddr);
    static fromString(endpoint: string): BuckyResult<Endpoint>;
    toString(): string;
    flag(): number;
    raw_measure(ctx?: any): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
// ...
export type Option<T> = (SomeOption<T> | NoneOption) & BaseOption<T>;
export function Some<T>(val: T): Option<T>;
export const None: NoneOption;
export class OptionEncoder<T extends RawEncode> implements RawEncode {
    val: Option<T>;
    constructor(val: Option<T>);
    value(): Option<T>;
    static from<V extends RawEncode, V1>(val: Option<V1>, ve: (v: V1) => V): OptionEncoder<V>;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class OptionWrapper<T extends RawEncode> implements RawEncode {
    constructor(v: Option<T>);
    value(): Option<T>;
    to<V1>(ve: (v: T) => V1): Option<V1>;
    raw_measure(ctx?: any): BuckyResult<number>;
    raw_encode(buf: Uint8Array, ctx?: any): BuckyResult<Uint8Array>;
}
export class OptionDecoder<T extends RawEncode> implements RawDecode<OptionWrapper<T>> {
    readonly decoder: RawDecode<T>;
    constructor(decoder: RawDecode<T>);
    raw_decode(buf: Uint8Array): BuckyResult<[OptionWrapper<T>, Uint8Array]>;
}
// ...
export enum BuckyErrorCode {
    Ok = 0,
    Failed = 1,
    InvalidParam = 2,
    Timeout = 3,
    NotFound = 4,
    AlreadyExists = 5,
    NotSupport = 6,
    ErrorState = 7,
    InvalidFormat = 8,
    Expired = 9,
    OutOfLimit = 10,
    InternalError = 11,
    PermissionDenied = 12,
    ConnectionRefused = 13,
    ConnectionReset = 14,
    ConnectionAborted = 15,
    NotConnected = 16,
    AddrInUse = 18,
    AddrNotAvailable = 19,
    Interrupted = 20,
    InvalidInput = 21,
    InvalidData = 22,
    WriteZero = 23,
    UnexpectedEof = 24,
    BrokenPipe = 25,
    WouldBlock = 26,
    UnSupport = 27,
    Unmatch = 28,
    ExecuteError = 29,
    Reject = 30,
    Ignored = 31,
    InvalidSignature = 32,
    AlreadyExistsAndSignatureMerged = 33,
    ConnectFailed = 40,
    ConnectInterZoneFailed = 41,
    InnerPathNotFound = 42,
    Conflict = 50,
    MongoDBError = 99,
    SqliteError = 100,
    UrlError = 101,
    ZipError = 102,
    HttpError = 103,
    JsonError = 104,
    HexError = 105,
    RsaError = 106,
    CryptoError = 107,
    MpscSendError = 108,
    MpscRecvError = 109,
    IoError = 110,
    NetworkError = 111,
    CodeError = 250,
    UnknownBdtError = 253,
    UnknownIOError = 254,
    Unknown = 255,
    Pending = 256,
    NotChange = 257,
    NotMatch = 258,
    NotImplement = 259,
    NotInit = 260,
    ParseError = 261,
    NotHandled = 262,
    MetaError = 5000,
    DecError = 15000
}
// ...
export class BuckyError {
    readonly m_code: BuckyErrorCodeEx;
    readonly m_msg: string;
    readonly m_origin?: string;
    constructor(code: number | string | BuckyErrorCodeEx, msg: string, origin?: string);
    static new_meta_error(meta_err: number, msg: string, origin?: string): BuckyError;
    static new_dec_error(dec_err: number, msg: string, origin?: string): BuckyError;
    static from(arg: string | BuckyErrorCode): BuckyError;
    get code(): BuckyErrorCode;
    get code_ex(): BuckyErrorCodeEx;
    is_system_error(): boolean;
    is_meta_error(): boolean;
    is_dec_error(): boolean;
    get value(): number;
    get msg(): string;
    get origin(): string | undefined;
    format(): string;
    error_with_log<T>(msg: string): BuckyResult<T>;
    toString(): string;
}
export type BuckyResult<T> = Result<T, BuckyError>;

export function bucky_time_now(): JSBI;
export function bucky_time(date: Date): JSBI;
export function bucky_time_2_js_time(val: JSBI): number;
export function bucky_time_2_js_date(val: JSBI): Date;
export function js_time_to_bucky_time(val: number): JSBI;
export function js_date_to_bucky_time(date: Date): JSBI;
export function sleep(ms: number): Promise<unknown>;

export class Vec<T extends RawEncode> implements RawEncode {
    constructor(val: T[]);
    value(): T[];
    static from<V extends RawEncode, V1>(val: V1[], ve: (v: V1) => V): Vec<V>;
    to<V1>(ve: (v: T) => V1): V1[];
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
}
export class VecDecoder<T extends RawEncode> implements RawDecode<Vec<T>> {
    readonly size_decoder: BuckySizeDecoder;
    readonly decoder: RawDecode<T>;
    constructor(decoder: RawDecode<T>);
    raw_decode(buf: Uint8Array): BuckyResult<[Vec<T>, Uint8Array]>;
}
// ...
export interface SubDescType {
        owner_type: "disable" | "option";
        area_type: "disable" | "option";
        author_type: "disable" | "option";
        key_type: "disable" | "single_key" | "mn_key" | "any";
}
export class ContentCodecInfo {
        version: number;
        format: number;
        constructor(version: number, format: number);
        static default(): ContentCodecInfo;
}
export abstract class DescTypeInfo {
        get_sub_obj_type(): number;
        set_sub_obj_type(v: number): void;
        abstract obj_type(): number;
        abstract sub_desc_type(): SubDescType;
}
export abstract class DescContent implements RawEncode {
        abstract type_info(): DescTypeInfo;
        codec_info(): ContentCodecInfo;
        abstract raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        abstract raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
}
export abstract class DescContentDecoder<T extends DescContent> implements RawDecode<T> {
        abstract type_info(): DescTypeInfo;
        abstract raw_decode(buf: Uint8Array, ctx: ContentRawDecodeContext): BuckyResult<[T, Uint8Array]>;
}
// ...
export abstract class BodyContent implements RawEncode {
        codec_info(): ContentCodecInfo;
        abstract raw_measure(ctx?: any, purpose?: RawEncodePurpose): BuckyResult<number>;
        abstract raw_encode(buf: Uint8Array, ctx?: any, purpose?: RawEncodePurpose): BuckyResult<Uint8Array>;
}
export abstract class BodyContentDecoder<T extends BodyContent> implements RawDecode<T> {
        abstract raw_decode(buf: Uint8Array, ctx: ContentRawDecodeContext): BuckyResult<[T, Uint8Array]>;
}
// ...
export class ObjectId implements RawEncode, Compareable<ObjectId> {
    m_buf: Uint8Array;
    m_base58?: string;
    constructor(buf: Uint8Array);
    get object_id(): void;
    static default(): ObjectId;
    static copy_from_slice(buf: Uint8Array): ObjectId;
    obj_type_code(): ObjectTypeCode;
    raw_measure(): BuckyResult<number>;
    raw_encode(buf: Uint8Array): BuckyResult<Uint8Array>;
    as_slice(): Uint8Array;
    clone(): ObjectId;
    length(): number;
    toString(): string;
    toJSON(): string;
    to_string(): string;
    to_base_58(): string;
    hashCode(): symbol;
    equals(other: ObjectId): boolean;
    eq(other: ObjectId): boolean;
    static from_str(s: string): BuckyResult<ObjectId>;
    static from_base_58(s: string): BuckyResult<ObjectId>;
    to_hash_value(): HashValue;
    object_category(): ObjectCategory;
    is_stand_object(): boolean;
    is_core_object(): boolean;
    is_dec_app_object(): boolean;
    info(): ObjectIdInfo;
}
export class ObjectIdDecoder implements RawDecode<ObjectId> {
    raw_decode(buf: Uint8Array): BuckyResult<[ObjectId, Uint8Array]>;
}
// ...
export enum ObjectTypeCode {
    Device = 1,
    People = 2,
    SimpleGroup = 3,
    Org = 4,
    AppGroup = 5,
    UnionAccount = 6,
    Chunk = 7,
    File = 8,
    Dir = 9,
    Diff = 10,
    ProofOfService = 11,
    Tx = 12,
    Action = 13,
    ObjectMap = 14,
    Contract = 15,
    Custom = 16
}